Fix async worker insert at max delay expiration
The async worker is supposed to insert all enqueued rows if
maxDelay time has elapsed since the last insertion and fewer
than maxRows have been enqueued in the meanwhile.

However, the maxDelay timer is reset each time a new row is enqueued,
rather than each time a batch of rows is inserted, causing to the insertion
at maxDelay expiration to never occur.

As an example, if max delay is set to 10 sec. and maxRows is set to 500
and a row is enqueued every 9 sec., then insertions will always occur
each 500*9 sec. rather than each 10 sec. as expected.

This commit fixes this issue.