Fix issues with insert retries
This commit fixes three distinct issues with the insertion retry
mechanism.

First, a sync worker used to retry only maxRetries - 1 times rather than
the expected maxRetries. As a result, if a worker was configured with
maxRetry = 0, which is a valid configuration value, then no insertion
would have ever succeeded.

Second, if a sync worker already tried the maximum number of times, it
would still wait retryInterval before returning. With this fix it
returns immediately if all allowed retries have been configured.

Third, a sync worker re-tried an insert only if the Google API replied
with a status code of 500 or 503. This commit makes sure that insertions
are retried also when they fail as a result of a network error, like a
transient loss of network connectivity.